BookRec/requirements.txt
streamlit==1.30.0
neo4j==5.14.0
qdrant-client==1.7.0
python-dotenv==1.0.0
pandas==2.1.3
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.30.2